refactor: verification 패키지 분리
